Resolved conflicts:
	pact/pact-clients/src/main/java/eu/stratosphere/pact/client/nephele/api/Client.java
	pact/pact-clients/src/main/java/eu/stratosphere/pact/client/nephele/api/PactProgram.java

+++ nephele/nephele-visualization/src/main/java/eu/stratosphere/nephele/visualization/swt/InstanceVisualizationData.java
@@ -181,7 +181,7 @@ public class InstanceVisualizationData {
 
 	private final double toMBitPerSec(long numberOfBytes, long profilingPeriod) {
 
-		return ((double) numberOfBytes / (BYTE_TO_MEGABIT * profilingPeriod / 1000));
+		return (((double) numberOfBytes) / ((double) (BYTE_TO_MEGABIT * profilingPeriod / 1000L)));
 	}
 
 	public boolean isProfilingEnabledForJob() {
